Fix talkTimePercentage type: Change from Int to Double
PROBLEM:
- Backend sends talkTimePercentage as Double (28.5)
- iOS model defined as Int
- Decode error: "Parsed JSON number <28.5> does not fit in Int"

SOLUTION:
- Change talkTimePercentage type from Int to Double in SpeakerMood struct
- Update preview mock data to use Double values

CHANGES:
- Models/Recording.swift: talkTimePercentage: Int → Double
- Models.swift: Same fix for consistency
- RecordingDetailView.swift: Update mock data to use 45.0 instead of 45

TESTING:
- iOS can now decode speaker moods with decimal percentages
- Firestore listener completes successfully
- Full async audio processing flow works end-to-end

+++ Models/Recording.swift
@@ -154,7 +154,7 @@ struct SpeakerMood: Codable, Identifiable {
     let speaker: String
     let mood: String // positive, neutral, negative
     let energy: Int // 0-100
-    let talkTimePercentage: Int
+    let talkTimePercentage: Double // Backend sends decimal values like 28.5
 
     enum CodingKeys: String, CodingKey {
         case speaker, mood, energy, talkTimePercentage
